+++ tests/e2e/auth/keeper/module_test.go
@@ -1,4 +1,4 @@
-package auth_test
+package keeper
 
 import (
 	"testing"
@@ -8,7 +8,6 @@ import (
 	"cosmossdk.io/depinject"
 	"cosmossdk.io/log"
 	"cosmossdk.io/x/auth/keeper"
-	"cosmossdk.io/x/auth/testutil"
 	"cosmossdk.io/x/auth/types"
 
 	simtestutil "github.com/cosmos/cosmos-sdk/testutil/sims"
@@ -18,7 +17,7 @@ func TestItCreatesModuleAccountOnInitBlock(t *testing.T) {
 	var accountKeeper keeper.AccountKeeper
 	app, err := simtestutil.SetupAtGenesis(
 		depinject.Configs(
-			testutil.AppConfig,
+			AppConfig,
 			depinject.Supply(log.NewNopLogger()),
 		),
 		&accountKeeper)