+++ x/gov/common_test.go
@@ -2,7 +2,6 @@ package gov_test
 
 import (
 	"cosmossdk.io/math"
-	_ "cosmossdk.io/x/consensus"
 	"cosmossdk.io/x/gov/types/v1beta1"
 	stakingtypes "cosmossdk.io/x/staking/types"
 )